python/eg_cloud_foundry_mongodb_hack/requirements.txt
pymongo==3.4.0
mongoengine==0.11.0
flask-mongoengine==0.9.3
Flask-RESTful==0.3.5
bcrypt==3.1.3
